+++ src/main/scala/com/ucesys/sparkscope/SparkScopeApp.scala
@@ -18,10 +18,11 @@
 package com.ucesys.sparkscope
 
 import com.ucesys.sparkscope.common.SparkScopeLogger
-import com.ucesys.sparkscope.io.file.FileReaderFactory
+import com.ucesys.sparkscope.io.reader.FileReaderFactory
 import com.ucesys.sparkscope.io.metrics.{MetricReaderFactory, MetricsLoaderFactory}
 import com.ucesys.sparkscope.io.property.PropertiesLoaderFactory
-import com.ucesys.sparkscope.view.ReportGeneratorFactory
+import com.ucesys.sparkscope.io.report.ReporterFactory
+import com.ucesys.sparkscope.io.writer.FileWriterFactory
 import org.apache.spark.SparkConf
 
 object SparkScopeApp {
@@ -37,7 +38,8 @@ object SparkScopeApp {
             fileReaderFactory = new FileReaderFactory(parsedArgs.region),
             propertiesLoaderFactory = new PropertiesLoaderFactory,
             metricsLoaderFactory = new MetricsLoaderFactory(new MetricReaderFactory(offline = true)),
-            reportGeneratorFactory = new ReportGeneratorFactory,
+            fileWriterFactory = new FileWriterFactory,
+            reportGeneratorFactory = new ReporterFactory,
         )
     }
 
@@ -47,13 +49,15 @@ object SparkScopeApp {
                         fileReaderFactory: FileReaderFactory,
                         propertiesLoaderFactory: PropertiesLoaderFactory,
                         metricsLoaderFactory: MetricsLoaderFactory,
-                        reportGeneratorFactory: ReportGeneratorFactory)
+                        fileWriterFactory: FileWriterFactory,
+                        reportGeneratorFactory: ReporterFactory)
                        (implicit logger: SparkScopeLogger): Unit = {
         val sparkScopeRunner = new SparkScopeRunner(
             sparkScopeConfLoader,
             sparkScopeAnalyzer,
             propertiesLoaderFactory,
             metricsLoaderFactory,
+            fileWriterFactory,
             reportGeneratorFactory
         )