+++ src/main/scala/com/ucesys/sparkscope/SparkScopeJobListener.scala
@@ -59,7 +59,7 @@ class SparkScopeJobListener(var sparkConf: SparkConf, val runner: SparkScopeRunn
     }
 
     override def onTaskEnd(end: SparkListenerTaskEnd): Unit = {
-        taskAggMetrics.aggregate(end.taskMetrics, end.taskInfo)
+        taskAggMetrics.aggregate(Option(end.taskMetrics), Option(end.taskInfo))
     }
 
     override def onApplicationStart(applicationStart: SparkListenerApplicationStart): Unit = {
@@ -107,11 +107,12 @@ class SparkScopeJobListener(var sparkConf: SparkConf, val runner: SparkScopeRunn
 
     def runSparkScopeAnalysis(applicationEnd: Option[Long]) = {
         val appStartEvent = applicationStartEvent.getOrElse(throw new IllegalArgumentException("App start event is empty"))
-
         val appContext = AppContext(
             appId = appStartEvent.appId.getOrElse(throw new IllegalArgumentException("App Id is empty")),
+            appName = appStartEvent.appName,
             appStartTime = appStartEvent.time,
             appEndTime = applicationEnd,
+            driverHost = sparkConf.getOption("spark.driver.host"),
             executorMap = executorMap.toMap,
             stages = stageMap.values.toSeq
         )