+++ src/main/scala/com/ucesys/sparkscope/SparkScopeRunner.scala
@@ -22,26 +22,29 @@ import com.ucesys.sparkscope.SparkScopeRunner.SparkScopeSign
 import com.ucesys.sparkscope.agg.TaskAggMetrics
 import com.ucesys.sparkscope.io.metrics.{MetricReaderFactory, MetricsLoaderFactory}
 import com.ucesys.sparkscope.io.property.PropertiesLoaderFactory
+import com.ucesys.sparkscope.io.report.{Reporter, ReporterFactory}
+import com.ucesys.sparkscope.io.writer.FileWriterFactory
 import com.ucesys.sparkscope.metrics.SparkScopeResult
-import com.ucesys.sparkscope.view.ReportGeneratorFactory
 import org.apache.spark.SparkConf
 
 import java.io.FileNotFoundException
-import java.nio.file.NoSuchFileException
+import java.nio.file.{NoSuchFileException, Paths}
 
 class SparkScopeRunner(sparkScopeConfLoader: SparkScopeConfLoader,
                        sparkScopeAnalyzer: SparkScopeAnalyzer,
                        propertiesLoaderFactory: PropertiesLoaderFactory,
                        metricsLoaderFactory: MetricsLoaderFactory,
-                       reportGeneratorFactory: ReportGeneratorFactory)
+                       fileWriterFactory: FileWriterFactory,
+                       reporterFactory: ReporterFactory)
                       (implicit val logger: SparkScopeLogger) {
     def run(appContext: AppContext, sparkConf: SparkConf, taskAggMetrics: TaskAggMetrics): Unit = {
         logger.info(SparkScopeSign, this.getClass)
 
         val sparkScopeStart = System.currentTimeMillis()
 
+        val sparkScopeConf = sparkScopeConfLoader.load(sparkConf, propertiesLoaderFactory)
+
         try {
-            val sparkScopeConf = sparkScopeConfLoader.load(sparkConf, propertiesLoaderFactory)
             val sparkScopeResult = this.runAnalysis(sparkScopeConf, appContext, taskAggMetrics)
 
             logger.info(s"${sparkScopeResult.stats.executorStats}\n", this.getClass)
@@ -49,7 +52,8 @@ class SparkScopeRunner(sparkScopeConfLoader: SparkScopeConfLoader,
             logger.info(s"${sparkScopeResult.stats.clusterMemoryStats}\n", this.getClass)
             logger.info(s"${sparkScopeResult.stats.clusterCPUStats}\n", this.getClass)
 
-            reportGeneratorFactory.get(sparkScopeConf).generate(sparkScopeResult)
+            val reporters: Seq[Reporter] = reporterFactory.get(appContext, sparkScopeConf)
+            reporters.foreach(_.report(sparkScopeResult))
         } catch {
             case ex: FileNotFoundException => logger.error(s"SparkScope couldn't open a file. SparkScope will now exit.", ex, this.getClass)
             case ex: NoSuchFileException => logger.error(s"SparkScope couldn't open a file. SparkScope will now exit.", ex, this.getClass)
@@ -58,13 +62,18 @@ class SparkScopeRunner(sparkScopeConfLoader: SparkScopeConfLoader,
         } finally {
             val durationSparkScope = (System.currentTimeMillis() - sparkScopeStart) * 1f / 1000f
             logger.info(s"SparkScope analysis took ${durationSparkScope}s", this.getClass)
+
+            val logFileWriter = fileWriterFactory.get(sparkScopeConf.logPath, sparkScopeConf.region)
+            val logPath = Paths.get(sparkScopeConf.logPath, s"${appContext.appId}.log")
+            logFileWriter.write(logPath.toString, logger.toString)
+            logger.info(s"Log saved to ${logPath}", this.getClass)
         }
     }
 
     def runAnalysis(sparkScopeConf: SparkScopeConf, appContext: AppContext, taskAggMetrics: TaskAggMetrics): SparkScopeResult = {
         val metricsLoader = metricsLoaderFactory.get(sparkScopeConf, appContext)
         val driverExecutorMetrics = metricsLoader.load(appContext, sparkScopeConf)
-        sparkScopeAnalyzer.analyze(driverExecutorMetrics, appContext, taskAggMetrics)
+        sparkScopeAnalyzer.analyze(driverExecutorMetrics, appContext, sparkScopeConf, taskAggMetrics)
     }
 }
 
@@ -75,7 +84,7 @@ object SparkScopeRunner {
           |    / __/__  ___ _____/ /__ / __/_ ___  ___  ___
           |   _\ \/ _ \/ _ `/ __/  '_/_\ \/_ / _ \/ _ \/__/
           |  /___/ .__/\_,_/_/ /_/\_\/___/\__\_,_/ .__/\___/
-          |     /_/                             /_/    spark3-v0.1.8
+          |     /_/                             /_/    spark3-v0.1.9
           |""".stripMargin
 
     def apply(): SparkScopeRunner = {
@@ -85,7 +94,8 @@ object SparkScopeRunner {
             new SparkScopeAnalyzer,
             new PropertiesLoaderFactory,
             new MetricsLoaderFactory(new MetricReaderFactory(offline = false)),
-            new ReportGeneratorFactory
+            new FileWriterFactory,
+            new ReporterFactory
         )
     }
 }