+++ src/main/scala/com/ucesys/sparkscope/io/FSPrefixes.scala
@@ -1,0 +1,6 @@
+package com.ucesys.sparkscope.io
+
+object FSPrefixes {
+    val HadoopFSPrefixes = Seq("maprfs", "hdfs", "file")
+    val S3Prefix = "s3"
+}