+++ src/main/scala/com/ucesys/sparkscope/io/file/FSPrefixes.scala
@@ -1,6 +1,0 @@
-package com.ucesys.sparkscope.io.file
-
-object FSPrefixes {
-    val HadoopFSPrefixes = Seq("maprfs", "hdfs", "file")
-    val S3Prefix = "s3"
-}