+++ src/main/scala/com/ucesys/sparkscope/io/metrics/S3OfflineMetricReader.scala
@@ -3,7 +3,7 @@ package com.ucesys.sparkscope.io.metrics
 import com.amazonaws.services.s3.{AmazonS3, AmazonS3ClientBuilder}
 import com.ucesys.sparkscope.common.{MetricType, SparkScopeConf, AppContext, SparkScopeLogger}
 import com.ucesys.sparkscope.data.DataTable
-import com.ucesys.sparkscope.io.file.S3FileReader
+import com.ucesys.sparkscope.io.reader.S3FileReader
 
 import java.nio.file.Paths