+++ src/main/scala/com/ucesys/sparkscope/io/reader/FileReader.scala
@@ -1,0 +1,5 @@
+package com.ucesys.sparkscope.io.reader
+
+trait FileReader {
+    def read(pathStr: String): String
+}