+++ src/main/scala/com/ucesys/sparkscope/io/report/DiagnosticsReport.scala
@@ -1,0 +1,6 @@
+package com.ucesys.sparkscope.io.report
+
+import com.ucesys.sparkscope.common.AppContext
+import com.ucesys.sparkscope.stats.SparkScopeStats
+
+case class DiagnosticsReport(appContext: AppContext, stats: SparkScopeStats)