+++ src/main/scala/com/ucesys/sparkscope/io/writer/TextFileWriter.scala
@@ -1,0 +1,7 @@
+package com.ucesys.sparkscope.io.writer
+
+trait TextFileWriter {
+    def write(destination: String, content: String): Unit
+
+    def exists(path: String): Boolean
+}