+++ src/main/scala/com/ucesys/sparkscope/view/warning/Warning.scala
@@ -1,0 +1,5 @@
+package com.ucesys.sparkscope.view.warning
+
+trait Warning {
+    def toString: String
+}