+++ src/main/scala/com/ucesys/sparkscope/warning/Warning.scala
@@ -1,5 +1,0 @@
-package com.ucesys.sparkscope.warning
-
-trait Warning {
-    def toString: String
-}