+++ src/test/scala/com/ucesys/sparkscope/io/metrics/CsvHadoopMetricsLoaderSuite.scala
@@ -27,37 +27,11 @@ import org.scalatest.{FunSuite, GivenWhenThen}
 class CsvHadoopMetricsLoaderSuite extends FunSuite with MockFactory with GivenWhenThen {
     implicit val logger: SparkScopeLogger = stub[SparkScopeLogger]
 
-//    test("Incorrect csv files test") {
-//        Given("Some csv metrics for driver and executor contain more rows than others")
-//        val csvReaderMock = stub[HadoopMetricReader]
-//        val appContext = mockAppContext("csv-loader-incorrect-csvs")
-//        mockIncorrectDriverMetrics(csvReaderMock, appContext.appId)
-//        val metricsLoader = new CsvMetricsLoader(csvReaderMock)
-//
-//        When("loading metrics")
-//        val driverExecutorMetrics = metricsLoader.load(appContext, sparkScopeConf)
-//
-//        Then("Driver and Executor Metrics should be loaded")
-//        assert(driverExecutorMetrics.driverMetrics.length == DriverCsvMetrics.length)
-//        assert(driverExecutorMetrics.executorMetricsMap.size == 1)
-//        assert(driverExecutorMetrics.executorMetricsMap("1").length == ExecutorCsvMetrics.length)
-//
-//        And("CsvHadoopMetricsLoader should ignore extra rows for driver metrics")
-//        driverExecutorMetrics.driverMetrics.foreach { metric =>
-//            assert(metric.numRows == 12)
-//        }
-//
-//        And("CsvHadoopMetricsLoader should ignore extra rows for executor metrics")
-//        driverExecutorMetrics.executorMetricsMap("1").foreach { metric =>
-//            assert(metric.numRows == 10)
-//        }
-//    }
-
     test("Successful metrics load test") {
         Given("Correctly configured metrics properties path")
         And("Correct csv files")
         val csvReaderMock = stub[HadoopMetricReader]
-        val appContext = mockAppContext("csv-loader-successful")
+        val appContext = mockAppContext(SampleAppId, "csv-loader-successful")
         mockcorrectMetrics(csvReaderMock, appContext.appId)
         val metricsLoader = new CsvMetricsLoader(csvReaderMock)
 
@@ -81,7 +55,7 @@ class CsvHadoopMetricsLoaderSuite extends FunSuite with MockFactory with GivenWh
         Given("Correctly configured metrics properties path")
         And("Csv metrics for 4 out of 5 executors(metrics for last executor are missing)")
         val csvReaderMock = stub[HadoopMetricReader]
-        val appContext = mockAppContextMissingExecutorMetrics("csv-loader-missing-metrics")
+        val appContext = mockAppContextMissingExecutorMetrics(SampleAppId, "csv-loader-missing-metrics")
         mockcorrectMetrics(csvReaderMock, appContext.appId)
         val metricsLoader = new CsvMetricsLoader(csvReaderMock)